backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djoser==2.2.0
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.0
djoser==2.2.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2